feat: Add utility function for merging CSS class names
This commit adds a new utility function `cn` to the `utils.ts` file. The `cn` function takes in multiple CSS class names as arguments and uses the `clsx` and `tailwind-merge` libraries to merge and format them. This utility function will be useful for combining and applying CSS classes in a more flexible and efficient way.