feat: Add name field to RegisterSchema for user registration
This commit adds a new field `name` to the `RegisterSchema` in the `schemas/index.ts` file. The `name` field is a string that must be at least 2 characters long. This enhancement allows users to provide their name during the registration process, improving the user experience and personalization of the application.