refactor: Update Jest configuration to generate JUnit test results
This commit modifies the `jest.config.ts` file to include the `jest-junit` reporter in the Jest configuration. The reporter is configured to output test results in JUnit XML format and save them in the `test-results` directory with the name `results.xml`. This change enables the generation of JUnit test reports, which can be used for further analysis and integration with other tools.