feat: Add generatePasswordResetToken function for password reset
This commit adds the generatePasswordResetToken function to the tokens.ts file. The function generates a password reset token for a given email address. If there is an existing token for the email, it is deleted before generating a new one. The token has an expiration time of 1 hour and is stored in the passwordResetToken table in the database.