feat: Add two-factor authentication support
This commit adds support for two-factor authentication. It introduces a new model called `TwoFactorConfirmation` to store the confirmation status of two-factor authentication for each user. The `User` model now has a new field `isTwoFactorEnabled` to indicate whether two-factor authentication is enabled for the user. Additionally, a new model `TwoFactorToken` is added to store the two-factor authentication tokens. This feature enhances the security of user accounts by providing an additional layer of authentication.